Sidebar to Dashboard page working, just need to populate dashboard pages in dashboard pages dir
