+++ src/Constants/role_sidebars.js
@@ -19,19 +19,8 @@ const role_sidebarinfo = {
   ]
 };
 
-const sidebar_to_comp = {
-  'My Re-medi Health Record': 'RemediRecord',
-  'Appointments': 'Appointments',
-  'Prescriptions': 'Prescriptions',
-  'Medical Records': 'Records',
-  'My Insurance Plan': 'InsurancePlan',
-  'Settings': 'Settings',
-  'Patients': 'Patients',
-  'Messages': 'Messages',
-  'Clients': 'Clients'
-};
+
 
 module.exports = {
-  role_sidebarinfo,
-  sidebar_to_comp
+  role_sidebarinfo
 };
