+++ src/Pages/Dashboard.js
@@ -5,9 +5,21 @@ import Footer from '../Components/Footer';
 
 import RemediRecord from './DashboardPages/RemediRecord';
 
-import { sidebar_to_comp } from '../Constants/role_sidebars';
+import './Dashboard.css';
+
 const role = "patient";
 
+const sidebar_to_comp = {
+  'My Re-medi Health Record': <RemediRecord/>,
+  'Appointments': 'Appointments',
+  'Prescriptions': 'Prescriptions',
+  'Medical Records': 'Records',
+  'My Insurance Plan': 'InsurancePlan',
+  'Settings': 'Settings',
+  'Patients': 'Patients',
+  'Messages': 'Messages',
+  'Clients': 'Clients'
+};
 export default class Dashboard extends Component{
   constructor(props) {
     super(props);
@@ -15,27 +27,28 @@ export default class Dashboard extends Component{
       dashpage: this.props.dashpage === null || this.props.dashpage === undefined ? null : this.props.dashpage
     }
     this.handleClick = this.handleClick.bind(this);
+    this.renderSwitch = this.renderSwitch.bind(this);
   }
 
   handleClick(item) {
     const dashpage = sidebar_to_comp[item];
     console.log(`CLICKED ${item} and rendering to ${sidebar_to_comp[item]}`);
-    this.setState({ dashpage: `<${dashpage}/>` });
+    this.setState({ dashpage: dashpage });
   }
+
+  renderSwitch(dashpage) { 
+    return dashpage === null ? "Let's fill out some health forms!" : dashpage
+  }
+
   render() {
     return (
-      <div style={{
-        display: 'flex',
-        flexDirection: 'column',
-        minHeight: '100vh'
-      }}>
+      <div className='dashboard'>
         <Header role={role} />
-        <div style={{
-          minHeight: '100px',
-          flex: "1"
-        }}>
-          <Sidebar role={role} onClick={this.handleClick}/>
-          <p>{this.state.dashpage === null ? "Let's fill out some health forms!" : this.props.dashpage}</p>
+        <div className='dash-content'>
+          <Sidebar role={role} onClick={this.handleClick} />
+          <div className='dash-page'>
+            {this.renderSwitch(this.state.dashpage)}
+          </div>
         </div>
         <Footer role={role} />
       </div>
